+++ packages/gatsby-theme-docs/src/components/cards.js
@@ -5,17 +5,26 @@ import { designSystem, ContentNotifications } from '@commercetools-docs/ui-kit';
 
 // Explanation about the following sizes:
 // https://github.com/commercetools/commercetools-docs-kit/pull/427#discussion_r425442556
+// the "inContentColumn" variation prevents only one card being in one row in regular content (except mobile)
 const cardNarrowMinWidth = '242px';
 const cardRegularMinWidth = '328px';
+const cardRegularMinWidthInContentColumn = '288px';
 const CardsContainer = styled.ul`
   list-style: none;
   display: grid;
   gap: ${designSystem.dimensions.spacings.m};
   grid-auto-columns: 1fr;
-  grid-template-columns: ${(props) =>
-    `repeat( auto-fill, minmax(${
-      props.narrow ? cardNarrowMinWidth : cardRegularMinWidth
-    }, 1fr)) `};
+  grid-template-columns: ${(props) => {
+    let columnMinWidth;
+    if (props.narrow) {
+      columnMinWidth = cardNarrowMinWidth;
+    } else {
+      columnMinWidth = props.fitContentColumn
+        ? cardRegularMinWidthInContentColumn
+        : cardRegularMinWidth;
+    }
+    return `repeat( auto-fill, minmax(${columnMinWidth}, 1fr))`;
+  }};
 `;
 
 const Cards = (props) => {
@@ -71,6 +80,7 @@ const Cards = (props) => {
 Cards.propTypes = {
   clickable: PropTypes.bool,
   narrow: PropTypes.bool,
+  fitContentColumn: PropTypes.bool,
   smallTitle: PropTypes.bool,
   children: PropTypes.node.isRequired,
 };