+++ websites/docs-smoke-test/CHANGELOG.md
@@ -1,6 +1,16 @@
 # @commercetools-website/docs-smoke-test
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`c18f1fa`](https://github.com/commercetools/commercetools-docs-kit/commit/c18f1fa7c31244912145fe4b8c8f5352fe1fd0ed), [`c18f1fa`](https://github.com/commercetools/commercetools-docs-kit/commit/c18f1fa7c31244912145fe4b8c8f5352fe1fd0ed)]:
+  - @commercetools-docs/gatsby-theme-docs@4.0.1
+  - @commercetools-docs/gatsby-theme-code-examples@3.0.2
+  - @commercetools-docs/gatsby-theme-constants@3.0.2
+
 ## 3.0.1
+
 ### Patch Changes
 
 - Updated dependencies [[`18976f7`](https://github.com/commercetools/commercetools-docs-kit/commit/18976f7344f0d5a219641da75ee996741e2d7ac2), [`f8d349b`](https://github.com/commercetools/commercetools-docs-kit/commit/f8d349b6a4f9d552fb66cdf96241b7f259c862d4)]: